// src/modules/make_modules/warehouse/transfer/transfer.interface.ts
import { Types } from "mongoose";

export type TStockTransfer = {
  _id?: Types.ObjectId;
  user_id: Types.ObjectId;
  product_id: Types.ObjectId;
  from_warehouse: Types.ObjectId;
  to_warehouse: Types.ObjectId;
  quantity: number;
  date: Date;
  notes?: string;
  createdAt?: Date;
  updatedAt?: Date;
};